requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-check-seo==0.3.4
django-classy-tags==1.0.0
django-cms==3.7.4
django-filer==1.7.1
django-formtools==2.2
django-js-asset==1.2.2
django-meta==1.7.0
django-mptt==0.11.0
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-link==2.6.1
djangocms-page-meta==0.10.0
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.9.1
djangocms-video==2.3.0
easy-thumbnails==2.7
html5lib==1.1
idna==2.10
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
Pillow==7.2.0
pyseoanalyzer==4.0.2
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.10
webencodings==0.5.1
